Dispatch MachineStateChangeCheckMessage after machine creation
