+++ src/Controller/JobController.php
@@ -5,6 +5,8 @@ namespace App\Controller;
 use App\Entity\Job;
 use App\Enum\ErrorResponseType;
 use App\Exception\EmptyUlidException;
+use App\Message\MachineStateChangeCheckMessage;
+use App\MessageDispatcher\MachineStateChangeCheckMessageDispatcher;
 use App\Repository\JobRepository;
 use App\Response\ErrorResponse;
 use App\Services\ErrorResponseFactory;
@@ -36,6 +38,7 @@ class JobController
         ResultsClient $resultsClient,
         ErrorResponseFactory $errorResponseFactory,
         WorkerManagerClient $workerManagerClient,
+        MachineStateChangeCheckMessageDispatcher $machineStateChangeCheckMessageDispatcher,
     ): JsonResponse {
         try {
             $id = $ulidFactory->create();
@@ -76,6 +79,10 @@ class JobController
 
         $job = new Job($id, $user->getUserIdentifier(), $suiteId, $resultsJob->token);
         $repository->add($job);
+
+        $machineStateChangeCheckMessageDispatcher->dispatch(
+            MachineStateChangeCheckMessage::createFromMachine($user->getSecurityToken(), $machine)
+        );
 
         return new JsonResponse([
             'job' => $job->jsonSerialize(),
