+++ tests/Application/AbstractCreateJobTest.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace App\Tests\Application;
 
-use App\Entity\Job;
-use App\Repository\JobRepository;
 use App\Tests\Services\AuthenticationConfiguration;
 use Symfony\Component\Uid\Ulid;
 
@@ -90,53 +88,5 @@ abstract class AbstractCreateJobTest extends AbstractApplicationTest
                 },
             ],
         ];
-    }
-
-    public function testCreateSuccess(): void
-    {
-        $jobRepository = self::getContainer()->get(JobRepository::class);
-        \assert($jobRepository instanceof JobRepository);
-        self::assertCount(0, $jobRepository->findAll());
-
-        $response = $this->applicationClient->makeCreateJobRequest(
-            self::$authenticationConfiguration->getValidApiToken(),
-            $this->suiteId,
-        );
-
-        self::assertSame(200, $response->getStatusCode());
-        self::assertSame('application/json', $response->getHeaderLine('content-type'));
-
-        $responseData = json_decode($response->getBody()->getContents(), true);
-        self::assertIsArray($responseData);
-
-        self::assertArrayHasKey('job', $responseData);
-        $jobData = $responseData['job'];
-
-        self::assertArrayHasKey('id', $jobData);
-        self::assertTrue(Ulid::isValid($jobData['id']));
-
-        self::assertArrayHasKey('suite_id', $jobData);
-        self::assertSame($this->suiteId, $jobData['suite_id']);
-
-        self::assertArrayHasKey('machine', $responseData);
-        $machineData = $responseData['machine'];
-
-        self::assertArrayHasKey('id', $machineData);
-        self::assertSame($jobData['id'], $machineData['id']);
-
-        self::assertArrayHasKey('state', $machineData);
-        self::assertSame('create/received', $machineData['state']);
-
-        self::assertArrayHasKey('ip_addresses', $machineData);
-        self::assertSame([], $machineData['ip_addresses']);
-
-        $jobs = $jobRepository->findAll();
-        self::assertCount(1, $jobs);
-
-        $job = $jobs[0];
-        self::assertInstanceOf(Job::class, $job);
-        self::assertSame($job->userId, self::$authenticationConfiguration->getUser()->id);
-        self::assertSame($job->suiteId, $this->suiteId);
-        self::assertNotNull($job->resultsToken);
     }
 }
