+++ tests/Functional/Application/CreateJobSuccessTest.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Tests\Functional\Application;
+
+use App\Tests\Application\AbstractCreateJobSuccessTest;
+
+class CreateJobSuccessTest extends AbstractCreateJobSuccessTest
+{
+    use GetClientAdapterTrait;
+}
